requirements.txt
langchain
langchain-google-genai
requests
python-dotenv
flask
matplotlib
pandas
geopy